fix the data format of those domains class

+++ src/main/java/com/yugao/constants/RedisKeyConstants.java
@@ -17,7 +17,7 @@ public final class RedisKeyConstants {
         return CAPTCHA_VERIFIED + username;
     }
 
-    public static String userIdAccessToken(Integer userId) {
+    public static String userIdAccessToken(Long userId) {
         return ACCESS_TOKEN + userId;
     }
 }
